backend/requirements.txt
fastapi
uvicorn
pydantic
GitPython
